Visual & Structural Foundation
The game has a structural foundation based off of a central Game class that has all of the basic funcctions of a game loop.

It also currectly has a visual foundation, based off of simulating a SNES pixel-art graphic style using SDL_gpu with a virtual resolution. It uses a viewport to make it to where no matter the window size, it keeps a 3:4 ratio centered on the screen.